requirements.txt
httpx[http2]>=0.27
pyyaml>=6.0
rich>=13.7
